+++ src/util.rs
@@ -1,5 +1,5 @@
 // defining a linear reltionship for testing
 
-pub fn linear(x : i64) -> i64 {
-    return 2*x;
+pub fn linear(x : f64) -> f64 {
+    return 2.0*x;
 }
